refactor(mobile): move TRPCProvider to lib folder
